back/requirements.txt
djangorestframework==2.3.13
pytz==2014.3
Django==1.6.5
psycopg2==2.5.2
South==1.0
